Speed up rendering for very large data
It does this by not rendering properties that aren't opened. It works
best when the data is chunked.

The test data has 100 chunks with 100 items each. Without this change,
it would take 2000ms to render all 10000 properties. With this change,
properties are re-rendered when clicked, which takes 30-50ms, which is a
huge improvement.

It's not perfect, but it's a huge improvement for viewing large amounts
of data with heavily nested properties.